requirements.txt
streamlit==1.32.0
langchain==0.1.20
langchain-google-genai==1.0.5
langchain-community==0.0.38
langchain-text-splitters==0.0.1
faiss-cpu==1.8.0
pypdf==3.17.4
python-dotenv==1.0.1